refactor: cleanup usage of next data handling to de-duplicate
